+++ NEWS.md
@@ -1,8 +1,13 @@
 # ympes development version
 
+- New function `fstrcapture()`. This is an (almost) drop-in replacement for
+  `strcapture()` that performs better when using Perl-compatible regexps.
+
 - **breaking change**. Assertion functions have been removed. They can still be
   found internally (and are checked by tests) but the intention is to use them
-  as a standalone file that is copy and pasted between packages as needed. 
+  as a standalone file that is copy and pasted between packages as needed.
+  
+- **Licence change**. Licence bumped from GPL-2 to GPL-3.
 
 # ympes 0.5.0